Create Paths.ts Utilities was getting a bit cluttered and we could stand to have all the path-related stuff in one place for clarity so I split these into their own file.
