+++ src/papyrus-lang-vscode/src/Utilities.ts
@@ -1,13 +1,12 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { PapyrusGame } from './PapyrusGame';
-
-import winreg from 'winreg';
 import { promisify } from 'util';
-import procList from 'ps-list';
 
-import { ExtensionContext, CancellationTokenSource } from 'vscode';
+import { CancellationTokenSource } from 'vscode';
+
+
 const exists = promisify(fs.exists);
+
 
 export function* flatten<T>(arrs: T[][]): IterableIterator<T> {
     for (const arr of arrs) {
@@ -17,78 +16,8 @@ export function* flatten<T>(arrs: T[][]): IterableIterator<T> {
     }
 }
 
-function getRegistryKeyForGame(game: PapyrusGame) {
-    switch (game) {
-        case PapyrusGame.fallout4:
-            return 'Fallout4';
-        case PapyrusGame.skyrim:
-            return 'Skyrim';
-        case PapyrusGame.skyrimSpecialEdition:
-            return 'Skyrim Special Edition';
-    }
-}
-
-export function getDevelopmentCompilerFolderForGame(game: PapyrusGame) {
-    switch (game) {
-        case PapyrusGame.fallout4:
-            return 'fallout4';
-        case PapyrusGame.skyrim:
-            return 'does-not-exist';
-        case PapyrusGame.skyrimSpecialEdition:
-            return 'skyrim';
-    }
-}
-
-export async function resolveInstallPath(
-    game: PapyrusGame,
-    installPath: string,
-    context: ExtensionContext
-): Promise<string> {
-    if (await exists(installPath)) {
-        return installPath;
-    }
-
-    const reg = new winreg({
-        key: `\\SOFTWARE\\${process.arch === 'x64' ? 'WOW6432Node\\' : ''}Bethesda Softworks\\${getRegistryKeyForGame(
-            game
-        )}`,
-    });
-
-    try {
-        const item = await promisify(reg.get).call(reg, 'installed path');
-
-        if (await exists(item.value)) {
-            return item.value;
-        }
-    } catch (_) { }
-
-    if (inDevelopmentEnvironment() && game !== PapyrusGame.skyrim) {
-        return context.asAbsolutePath('../../dependencies/compilers');
-    }
-
-    return null;
-}
-
 export function delayAsync(durationMs: number): Promise<void> {
     return new Promise((r) => setTimeout(r, durationMs));
-}
-
-export function getDefaultFlagsFileNameForGame(game: PapyrusGame) {
-    return game === PapyrusGame.fallout4 ? 'Institute_Papyrus_Flags.flg' : 'TESV_Papyrus_Flags.flg';
-}
-
-const executableNames = new Map([
-    [PapyrusGame.fallout4, 'Fallout4.exe'],
-    [PapyrusGame.skyrimSpecialEdition, 'SkyrimSE.exe'],
-]);
-
-export function getExecutableNameForGame(game: PapyrusGame) {
-    return executableNames.get(game);
-}
-
-export async function getGameIsRunning(game: PapyrusGame) {
-    const processList = await procList();
-    return processList.some((p) => p.name.toLowerCase() === getExecutableNameForGame(game).toLowerCase());
 }
 
 export async function waitWhile(
@@ -120,26 +49,6 @@ export function toCommandLineArgs(obj: Object): string[] {
             ];
         })
     );
-}
-
-function getToolGameName(game: PapyrusGame) {
-    switch (game) {
-        case PapyrusGame.fallout4:
-            return 'Fallout4';
-        case PapyrusGame.skyrim:
-        case PapyrusGame.skyrimSpecialEdition:
-            return 'Skyrim';
-    }
-}
-
-export function getLanguageToolPath(game: PapyrusGame) {
-    const toolGameName = getToolGameName(game);
-    return `./bin/Debug/net461/DarkId.Papyrus.Host.${toolGameName}/DarkId.Papyrus.Host.${toolGameName}.exe`;
-}
-
-export function getDebugToolPath(game: PapyrusGame) {
-    const toolGameName = getToolGameName(game);
-    return `./debug-bin/Debug/net461/DarkId.Papyrus.DebugAdapterProxy.${toolGameName}/DarkId.Papyrus.DebugAdapterProxy.${toolGameName}.exe`;
 }
 
 export function mkdirIfNeeded(pathname: string) {
